Add accent vector to inputs in chain model
This makes it possible to append arbitrary utterance-level vectors to
chain model inputs during training and parallel decoding on CPU. The
additional input will be labelled "accent".

Don't know if this would work if you tried to use it without also
supplying ivector inputs, and haven't tried anything with online
features -- this assumes a single additional vector per utterance which
is constant for the entire duration.